third_party/WebKit/LayoutTests/editing/selection/mouse/click-user-select-all-contenteditable.html

<!doctype HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<div id="log"></div>
<script>
test(function(){
    assert_true(window.hasOwnProperty('eventSender'), 'this test requires window.eventSender');
    assert_selection(
        '<div contenteditable="true" id="div" style="-webkit-user-select:all">bar</div>',
        selection => {
            var div = selection.document.getElementById('div');
            eventSender.mouseMoveTo(selection.document.offsetLeft + div.offsetLeft + 5, selection.document.offsetTop + div.offsetTop + 5);
            eventSender.mouseDown();
            eventSender.mouseUp();
            eventSender.leapForward(100);
        },
        '<div contenteditable="true" id="div" style="-webkit-user-select:all">|bar</div>');
    },
'User can focus in -webkit-user-select:all contenteditable element');
</script>
